Crate arceos_posix_api

POSIX-compatible APIs for ArceOS modules

Modules

config

Platform-specific constants and parameters.

ctypes

POSIX C types.

Statics

RX_ENVIRONalloc

Save environment variables

environalloc

A pointer pointing to RX_ENVIRON

Functions

environ_iteralloc

Generate an iterator for environment variables

sys_accept⚠net

Accept for connections on a socket

sys_bindnet

Bind a address to a socket.

sys_clock_gettime⚠

Get clock time since booting

sys_clock_settime⚠

Get clock time since booting

sys_clone⚠multitask and musl

Create new thread by sys_clone, return new thread ID

sys_closefd

Close a file by fd.

sys_connectnet

Connects the socket to the address specified.

sys_dupfd

Duplicate a file descriptor.

sys_dup2fd

Duplicate a file descriptor, but it uses the file descriptor number specified in new_fd.

sys_dup3fd and musl

dup3 used by A64 for MUSL

sys_epoll_createepoll and (select or poll or epoll)

Creates a new epoll instance.

sys_epoll_ctl⚠epoll and (select or poll or epoll)

Control interface for an epoll file descriptor

sys_epoll_pwait⚠epoll and (select or poll or epoll)

epoll_pwait used by A64. Currently ignore signals

sys_epoll_wait⚠epoll and (select or poll or epoll)

Waits for events on the epoll instance referred to by the file descriptor epfd.

sys_exit

Exit current task

sys_fcntlfd

Manipulate file descriptor.

sys_freeaddrinfo⚠net

Free queried addrinfo struct

sys_fstat⚠fs

retrieve information about the file pointed by fd

sys_fsync⚠fs

fsync

sys_futexmultitask

Futex implementation inspired by Starry

sys_getaddrinfo⚠net

Query addresses for a domain name.

sys_getcwdfs

Get the path of the current directory.

sys_getitimer⚠signal

Get timer to send signal after some time

sys_getpeername⚠net

Get peer address to which the socket sockfd is connected.

sys_getpid

Get current thread ID.

sys_getrlimit⚠

Get resource limitations

sys_getsockname⚠net

Get current address to which the socket sockfd is bound.

sys_invalid

Invalid syscall

sys_ioctlfd

ioctl implementation, currently only support fd = 1

sys_listennet

Listen for connections on a socket

sys_lseekfs

Set the position of the file indicated by fd.

sys_lstat⚠fs

Get the metadata of the symbolic link and write into buf.

sys_mkdirfs

Creates a new, empty directory at the provided path.

sys_mkdiratfs

attempts to create a directory named pathname under directory pointed by fd

sys_mmapalloc

Creates a new mapping in the virtual address space of the call‐ ing process.

sys_mprotectalloc

Changes the access protections for the calling process’s memory pages containing any part of the address range in the interval [addr, addr+len-1].
addr must be aligned to a page boundary.

sys_munmapalloc

Deletes the mappings for the specified address range

sys_nanosleep⚠

Sleep some nanoseconds

sys_newfstatat⚠fs

newfstatat used by A64

sys_openfs

Open a file by filename and insert it into the file descriptor table.

sys_openatfs

Open a file under a specific dir

sys_pipepipe

Create a pipe

sys_pipe2pipe

pipe2 syscall, used by AARCH64

sys_poll⚠poll and (select or poll or epoll)

Used to monitor multiple file descriptors for events

sys_ppoll⚠poll and (select or poll or epoll)

ppoll used by A64. Currently ignore signal

sys_prlimit64⚠

set/get resource limitations

sys_pselect6⚠select and (select or poll or epoll)

allows a program to monitor multiple file descriptors, waiting until one or more of the file descriptors become “ready” for some class of I/O operation (e.g., input possible)

sys_pthread_cond_broadcast⚠multitask

Restarts all the threads that are waiting on the condition variable.

sys_pthread_cond_destroy⚠multitask

Destroy a condition variable

sys_pthread_cond_init⚠multitask

Initialize a condition variable

sys_pthread_cond_signal⚠multitask

Restarts one of the threads that are waiting on the condition variable.

sys_pthread_cond_timedwait⚠multitask

Wait for the condition variable to be signaled or timeout

sys_pthread_cond_wait⚠multitask

Wait for the condition variable to be signaled

sys_pthread_create⚠multitask

Create a new thread with the given entry point and argument.

sys_pthread_exitmultitask

Exits the current thread. The value retval will be returned to the joiner.

sys_pthread_getspecificmusl and multitask

Get the value of a specific key for a thread.

sys_pthread_join⚠multitask

Waits for the given thread to exit, and stores the return value in retval.

sys_pthread_key_create⚠musl and multitask

Allocate a specific key for a process shared by all threads.

sys_pthread_key_deletemusl and multitask

Destroy a specific key for a process.

sys_pthread_mutex_destroymultitask

Destroy the given mutex.

sys_pthread_mutex_initmultitask

Initialize a mutex.

sys_pthread_mutex_lockmultitask

Lock the given mutex.

sys_pthread_mutex_trylockmultitask

Lock the given mutex like sys_pthread_mutex_lock, except that it does not block the calling thread if the mutex is already locked.

sys_pthread_mutex_unlockmultitask

Unlock the given mutex.

sys_pthread_selfmultitask

Returns the pthread struct of current thread.

sys_pthread_setspecificmusl and multitask

Set the value of a specific key for a thread.

sys_read

Read data from the file indicated by fd.

sys_recvnet

Receive a message on a socket.

sys_recvfrom⚠net

Receive a message on a socket and get its source address.

sys_renamefs

Rename old to new If new exists, it is first removed.

sys_rmdirfs

Remove a directory, which must be empty

sys_rt_sigaction

sigaction syscall for A64 musl

sys_rt_sigprocmask

Set mask for given thread

sys_sched_yield

Relinquish the CPU, and switches to another task.

sys_select⚠select and (select or poll or epoll)

Monitor multiple file descriptors, waiting until one or more of the file descriptors become “ready” for some class of I/O operation

sys_sendnet

Send a message on a socket to the address connected.

sys_sendmsg⚠net

Send a message on a socket to the address connected. The message is pointed to by the elements of the array msg.msg_iov.

sys_sendtonet

Send a message on a socket to the address specified.

sys_set_tid_addressmultitask and musl

Set child tid address

sys_setitimer⚠signal

Set a timer to send a signal to the current process after a specified time

sys_setrlimit⚠

Set resource limitations

sys_setsockoptnet

setsockopt, currently ignored

sys_shutdownnet

Shut down a full-duplex connection.

sys_sigactionsignal

Set signal handler

sys_socketnet

Create an socket for communication.

sys_stat⚠fs

Get the file metadata by path and write into buf.

sys_sysinfo⚠

Return sysinfo struct

sys_umask

umask

sys_unlinkfs

Removes a file from the filesystem.

sys_unlinkatfs

deletes a name from the filesystem

sys_write

Write data to the file indicated by fd.

sys_writev⚠

Write a vector.